sound mixer now operates on floating point, has on-the-fly resampling (allows features like pitch shifting, and potentially doppler), and the ogg and modplug streaming code has been largely rewritten, much cleaner now, this saves about 4MB of ram when playing quake content at default settings (by not resampling the wav data to 48khz anymore), snd_speed changes will now apply to next vid_restart (tested at 48, 96, 192khz)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@11352 d7cf8633-e32d-0410-b094-e92efae38249
::stable-branch::merge=ff0e0fd9ea6ce001e9c2c6f9e18ecda6aef15d10

+++ gl_textures.c
@@ -2208,7 +2208,7 @@ rtexture_t *R_LoadTextureDDSFile(rtexturepool_t *rtexturepool, const char *filen
 		bytesperblock = 8;
 		ddssize -= 128;
 		ddssize /= 2;
-		for (i = 0;i < ddssize;i += bytesperblock)
+		for (i = 0;i < (int)ddssize;i += bytesperblock)
 			memcpy(&ddspixels[i], &ddspixels[(i<<1)+8], 8);
 		ddssize += 128;
 	}
